Split webhook-price-signature tests into a folder and drop its max-lines exemption
The signed-price webhook oracle test file had grown to 1,263 lines and was
grandfathered into biome's noExcessiveLinesPerFile exemption list. Split it
into test/lib/webhook-price-signature/ — grouped by verdict/area (trusted &
mismatch, stored refunds & ignores, mismatch & package overrides, and hidden
& standalone members) — with the file's shared setup and assertion helpers
(and the package-setup helpers that lived mid-file) lifted into helpers.ts,
so every file sits under 400 lines, then removed the file's exemption entry.

The move surfaced two duplicated parent/child-standalone setups, now the
local helpers parentWithBookableAloneChild and foldedChildOrder. All 40
original tests carry over unchanged and pass; jscpd is back to 0%.
